### 2025-05-20 1dc6495
ensure that single-observation template can properly load RV from defined KW
