### 2025-05-30 #8239 @warrickball
Hi @Kamuish, thanks for your submission to JOSS and sorry it's taken so long to do the pre-review checks. All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Owing to our current backlog, it will probably be at least 3 months. Thanks for your patience and sorry for the further wait!

In the meantime, could you also comment on the development history of the code? It seems like quite a lot of code that was committed quite quickly. Was there perhaps development that happened before the initial commit? Were there specific code generation tools you used?
