### 2025-06-03 #8239 @warrickball
> Hi @warrickball , no worries on my side! I am guessing that this issue will be updated as soon as I have a new referee.

First one of our editors will assign themselves to the paper and they'll then start the process of finding (at least two) referees. But yes, you might not hear anything until that happens.

> Over time, new functionalities were added and I felt that it not longer made sense to have this framework tied to the RV extraction process. This separation was done (manually) over a short period of time (hence the short commit history of this repository).

Ah yes, I see what you mean. That's absolutely fine, and demonstrates that the codebase is better established than the short commit history might otherwise imply. Thanks!
