### 2025-06-05 cfa2379
Computation of BERV factor to bypass issues on ESO pipelines
