### 2025-06-05 9fe2c30
More explicit if/else clauses
