### 2025-06-05 0d6a3a7
Recomputation of BERV from BERV_FACTOR in ESO PIPE
We do not need to use the BERV factor, since we can recover the
BERV term from the BERV_FACTOR value in the header. Using the
BERV_factor will harden the expansion of the telluric template, since it
needs to know the exact value of the BERV (in the lines mode) and it
will not have access to BERV_factor
