### 2025-06-05 36451b6
Prepare BERV shift functions to use BERV factor
For now this will not be used, as it will only be relevant for ESO
pipelines (newer versions). However, in that case, we are recomputing
the BERV
