### 2025-06-05 37ae4c5
Log when BERV_factor is >10m/s different than berv
