### 2025-06-23 54f9166
general interface to custom fits files
