### 2025-07-02 b00e18d
adding BERV_FACTOR to ESO pipelines, after checking with J. Faria that this is correct
