### 2025-07-02 0561e52
re-adding option to iterate stellar templates from disk storage
