+++ docs/index.md
@@ -45,6 +45,7 @@ Management of any issues to the code, as well as external contribution is done t
 1. Addition of new interpolation algorithms
 2. Inclusion of interface to SPIROU simulated data
 3. Added BERV_factor to ESO pipelines
+4. Improvement of data interface to iterate over template construction
 
 ## V1.0.0 (5th of May 2025)
 
