### 2025-07-30 ce65c39
Add disk files for the loggers
