### 2025-08-05 cb88938
adding missing instrument from the general dict
