### 2025-08-05 5f2311d
add debug messages to stellar template
