### 2025-08-05 07ca0aa
Missing the rejection arrays from the single-OBS stellar template
