### 2025-08-05 c79487c
fixes to the CARMENES interface
