### 2025-08-29 #8239 @warrickball
Hi @Kamuish, I'm afraid not. Note that we first need an editor to pick this up for handling, and they'll then look for referees. You're into the top half of the queue at this point, and I'm expecting a few editors to return from vacation in the next few weeks, plus we should be recruiting more editors over the next few months. I do apologise for the long wait and can only say it'll probably still be another 2-3 months before someone becomes available. Sorry!
