### 2025-09-29 7189911
Use deterministic hash across runs
