### 2025-09-29 04a343d
fallback for lack of Moon info
