### 2025-09-29 1b2ecf7
Proper search for CCF files in HARPS data
