### 2025-09-29 0002728
fix number of bits in frameID
