### 2025-09-29 07115fb
Fix numpy version at installation
