+++ pyproject.toml
@@ -10,7 +10,7 @@ dependencies = [
     "h5py>=3.13.0",
     "loguru>=0.7.3",
     "matplotlib>=3.10.1",
-    "numpy<2.0.0",
+    "numpy>=1.23.2,<2.0.0",
     "pandas>=2.2.3",
     "scipy>=1.15.2",
     "sqlalchemy>=2.0.40",
@@ -21,7 +21,7 @@ dependencies = [
 ]
 
 [build-system]
-requires = ["setuptools",  "numpy"]
+requires = ["setuptools",  "numpy>=1.23.2,<2.0.0"]
 build-backend = "setuptools.build_meta"
 
 
