### 2025-10-20 b5d64f1
Oversampling of OBs for template construction
