### 2025-10-20 4ace869
Oversampling of OBs for template construction
