### 2025-10-20 aa640f7
Avoid imports for only type hints
