### 2025-10-20 df7cba2
automatically load array shape
