### 2025-10-22 9910284
Account for oversampled template size
