### 2025-10-30 #8239 @Kamuish
@warrickball I fully understand that the shortage of editors presents a significant challenge. However, my manuscript was submitted in May (over five months ago), and it has not yet been assigned to an editor. Each time I inquire about the status, I am informed that it will be a few more weeks.
