+++ src/ASTRA/data_objects/DataClass.py
@@ -18,6 +18,7 @@ from typing import (
 
 import numpy as np
 import ujson as json
+import hashlib
 from loguru import logger
 from tabletexifier import Table
 
@@ -62,6 +63,23 @@ def hash_path_to_int(fname: Path, algo: str = "sha256", bits: int = 32) -> int:
         return digest_int & mask
     return digest_int
 
+def hash_path_to_int(fname: Path, algo: str = "sha256", bits: int = 32) -> int:
+    """
+    Deterministic integer hash of a filesystem path.
+    - Uses hashlib (default sha256)
+    - Returns an integer, truncated to `bits` if specified
+    """
+
+    h = hashlib.new(algo)
+    h.update(fname.encode("utf-8"))
+    digest_int = int.from_bytes(h.digest(), "big")
+
+    if bits is not None:
+        mask = (1 << bits) - 1
+        return digest_int & mask
+    return digest_int
+
+
 class DataClass(BASE):
     """Standard interface to handle data.
 
@@ -84,7 +102,9 @@ class DataClass(BASE):
     This *data_object* has all the functions that the DataClass object implements!
     """
 
-    extra_loading_functions: Dict[str, Callable] = {"CARMENES": load_CARMENES_extra_information}
+    extra_loading_functions: Dict[str, Callable] = {
+        "CARMENES": load_CARMENES_extra_information
+    }
 
     def __init__(
         self,
@@ -139,7 +159,9 @@ class DataClass(BASE):
                     OBS_list.append(Path(line.split("\n")[0]))
 
         elif isinstance(input_files, Iterable):
-            logger.info("DataClass opening {} files from a list/tuple", len(input_files))
+            logger.info(
+                "DataClass opening {} files from a list/tuple", len(input_files)
+            )
 
             OBS_list = [Path(i) if isinstance(i, str) else i for i in input_files]
         else:
@@ -160,7 +182,8 @@ class DataClass(BASE):
                     instrument_options,
                     reject_subInstruments,
                     frameID=frameID,
-                    quiet_user_params=frameID != 0,  # Only the first frame will output logs
+                    quiet_user_params=frameID
+                    != 0,  # Only the first frame will output logs
                 ),
             )
 
@@ -181,7 +204,9 @@ class DataClass(BASE):
 
         # TODO: find a better way of doing this!
         self.Target = Target(
-            self.collect_KW_observations("OBJECT", self._inst_type.sub_instruments, include_invalid=False),
+            self.collect_KW_observations(
+                "OBJECT", self._inst_type.sub_instruments, include_invalid=False
+            ),
             original_name=target_name,
             target_dictionary_path=target_dictionary_path,
         )
@@ -220,15 +245,16 @@ class DataClass(BASE):
                 units=None,
                 apply_drift_corr=False,
             )
-
-            previous_filename = cube.cached_info["date_folders"][cube.frameIDs.index(frameID)]
-
+            previous_filename = cube.cached_info["date_folders"][
+                cube.frameIDs.index(frameID)
+            ]
             if previous_filename != frame.file_path:
                 msg = f"Loading RVs from cube with different frameID layouts of {frame.sub_instrument} ({previous_filename} vs {frame.file_path})"
                 logger.critical(msg)
                 raise InvalidConfiguration(msg)
 
             frame.store_previous_SBART_result(RV=sbart_rv, RV_err=sbart_uncert)
+        logger.info("Finished loading of previous s-BART RVs")
 
     def reject_order_region_from_frame(self, frameID: int, order: int, region):
         frame = self.get_frame_by_ID(frameID)
@@ -251,8 +277,12 @@ class DataClass(BASE):
         logger.info("Computing activity windows for each RV measurements")
         for frameID in self.get_valid_frameIDS():
             frame = self.get_frame_by_ID(frameID)
[…]
