### 2025-10-31 a833245
missing proper setup for CARMENES
