### 2025-10-31 ec0e349
Update logging strategy to avoid clash with sBART
