### 2025-11-03 58188fd
Update of Stellar Model to properly load previous RVs
