+++ src/ASTRA/Instruments/CARMENES.py
@@ -64,7 +64,9 @@ class CARMENES(Frame):
             description="If positive, then sigma clip the flux",
         ),
     )
-    _default_params.update("IS_SA_CORRECTED", UserParam(True, constraint=BooleanValue, mandatory=False))
+    _default_params.update(
+        "IS_SA_CORRECTED", UserParam(True, constraint=BooleanValue, mandatory=False)
+    )
 
     sub_instruments = {
         "CARMENES": datetime.datetime.max,
@@ -127,7 +129,9 @@ class CARMENES(Frame):
         # lat/lon from: https://geohack.toolforge.org/geohack.php?pagename=Calar_Alto_Observatory&params=37_13_25_N_2_32_46_W_type:landmark_region:ES
         # height from: https://en.wikipedia.org/wiki/Calar_Alto_Observatory
         lat, lon = 37.223611, -2.546111
-        self.instrument_properties["EarthLocation"] = EarthLocation.from_geodetic(lat=lat, lon=lon, height=2168)
+        self.instrument_properties["EarthLocation"] = EarthLocation.from_geodetic(
+            lat=lat, lon=lon, height=2168
+        )
 
         # from https://www.mide.com/air-pressure-at-altitude-calculator
         # and convert to Pa to mbar
@@ -149,7 +153,9 @@ class CARMENES(Frame):
 
         # Load BERV info + previous RV
         self.observation_info["MAX_BERV"] = 30 * kilometer_second
-        self.observation_info["BERV"] = header["HIERARCH CARACAL BERV "] * kilometer_second
+        self.observation_info["BERV"] = (
+            header["HIERARCH CARACAL BERV "] * kilometer_second
+        )
 
         # TODO: check ambient temperature on CARMENES data TO SEE IF IT IS THE "REAL ONE"
         # Environmental KWs for telfit (also needs airmassm previously loaded)
@@ -165,11 +171,17 @@ class CARMENES(Frame):
                     self.observation_info[name], old_scale="Celsius", new_scale="Kelvin"
                 )
         for order in range(self.N_orders):
-            self.observation_info["orderwise_SNRs"].append(header[f"HIERARCH CARACAL FOX SNR {order}"])
+            self.observation_info["orderwise_SNRs"].append(
+                header[f"HIERARCH CARACAL FOX SNR {order}"]
+            )
 
         try:
-            self.observation_info["MOON PHASE"] = header["HIERARCH CAHA INS SCHEDULER MOON PHASE"]
-            self.observation_info["MOON DISTANCE"] = header["HIERARCH CAHA INS SCHEDULER MOON DISTANCE"]
+            self.observation_info["MOON PHASE"] = header[
+                "HIERARCH CAHA INS SCHEDULER MOON PHASE"
+            ]
+            self.observation_info["MOON DISTANCE"] = header[
+                "HIERARCH CAHA INS SCHEDULER MOON DISTANCE"
+            ]
         except KeyError:
             self.observation_info["MOON PHASE"] = 0
             self.observation_info["MOON DISTANCE"] = 0
@@ -195,11 +207,15 @@ class CARMENES(Frame):
 
             night_drift = drift_file.split("s-")[0].split("car-")[-1]
             night_drift = datetime.datetime.strptime(night_drift, r"%Y%m%dT%Hh%Mm%S")
-            night_data = datetime.datetime.strptime(header["CARACAL DATE-OBS"], r"%Y-%m-%dT%H:%M:%S")
+            night_data = datetime.datetime.strptime(
+                header["CARACAL DATE-OBS"], r"%Y-%m-%dT%H:%M:%S"
+            )
             msg = f"\tDistance between calibration (FP) and OB has surpassed the limit of {self._internal_configs['max_hours_to_calibration']}"
             if (
                 abs(night_drift - night_data).total_seconds()
[…]
