### 2025-11-06 8eefb09
Avoid template construction test on CI
