+++ .github/workflows/testing.yml
@@ -24,4 +24,4 @@ jobs:
 
       - name: Run tests
         # For example, using `pytest`
-        run: uv run pytest tests
+        run: uv run python -m pytest -v tests/ -m "not slow"
