### 2025-11-09 #8239 @darthoctopus
I will be happy to review this paper, although I forecast delays in the next couple of weeks owing to some conference travel (for which I apologise).
