### 2025-11-14 #9413 @editorialbot
[REVIEW]: Astra: A Python Package for Cross-Instrument Stellar and Telluric Template Construction
**Submitting author:** @Kamuish (<a href="http://orcid.org/0000-0003-4920-738X">André Silva</a>)
**Repository:** https://github.com/Kamuish/ASTRA
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
