### 2025-11-16 #9413 @tnehezd
Hi @Kamuish! (It's not tagging you...)

I just need some assistance for the start. I read the paper, installed the package and checked your Git repo. I went through some py-s.

My questions are:

1. Do I need some spectra fits on my laptop to use your package? If so, that would be super helpful if you could upload a directory for some examples to test your package, or give some hint about where I can download some. Without that, I'ma afraid, I cannot move forward.
2. Second, as a side note: it would be also very useful, If the main page of your repo would have an "Installation" section, and a direct indication about the py version dependency. This is not only helpful for users, but also important because the JOSS reviewer checklist explicitly requires such information to be present.
3. I saw that you already provide some .md files and a number of Jupyter notebooks, which are all excellent resources — the notebooks in particular are very valuable! What I would recommend, however, is to make these materials more accessible by publishing them through a documentation platform such as Read the Docs. Even if the package itself is very useful, users may struggle if they cannot easily find clear, well‑structured documentation. Having a central documentation site would greatly improve usability and visibility.
4. I could not find a Community Guidelines section in the repository. While the existing materials are valuable, adding such a section would be highly beneficial. It helps set expectations for users and contributors, and it is explicitly emphasized in the JOSS reviewer checklist. Including clear community guidelines would therefore strengthen the submission and ensure compliance with the review criteria.

I'm opening issues for points 2-4, and wait for the answer for point 1.

Best regards,
Dóri
