### 2025-11-16 be7ea9f
Update installation instructions
