### 2025-11-16 15aa73f
add link to docs in readme file
