+++ README.md
@@ -2,9 +2,11 @@
 
 Welcome to the repository of ASTRA (Astrophysical Spectral Tools for Retrieval &amp; Analysis). 
 
-The main goal of this project is to standirdize the handling of stellar spectra for both RV extraction tools and spectral analysis tools. In order to do so, ASTRA provides:
+The main goal of this project is to standardize the handling of stellar spectra for both RV extraction tools and spectral analysis tools. In order to do so, ASTRA provides:
 
 1) Common interface to interact with stellar spectra of different state-of-the-art instruments
 2) Framework to construct stellar templates from data, following best practices
 3) Framework to mask telluric features from stellar spectra
 4) Implement utility methods to interpolate, normalize and smooth stellar spectra
+
+For information on installation and usage, check the official [documentation](https://kamuish.github.io/ASTRA/)
