### 2025-11-16 200eadd
update spectrograph documentation
