+++ docs/api/spectrographs.md
@@ -1,0 +1,7 @@
+# Currently available interfaces to spectrographs
+
+::: ASTRA.Instruments.ESPRESSO
+::: ASTRA.Instruments.HARPS
+::: ASTRA.Instruments.CARMENES
+::: ASTRA.Instruments.MAROONX
+
