+++ docs/api/Frame.md
@@ -1,9 +1,5 @@
 # Accessing stellar spectra
 
 ::: ASTRA.base_models.Frame
-
-# Currently available interfaces to spectrographs
-
-::: ASTRA.Instruments.ESPRESSO
-::: ASTRA.Instruments.HARPS
+::: ASTRA.Instruments.ESO_PIPELINE
 
