### 2025-11-17 b53024e
add info on contribution
