### 2025-11-17 f0e9d1e
add info on contribution
