### 2025-11-22 #9413 @tnehezd
Hey @Kamuish,

So I have some new remarks. I still struggle to launch ASTRA.

## Comment on ASTRA Package (Experience on macOS)

I tested the ASTRA package on macOS using Python 3.11 and 3.12 environments. I used both the PyPI version (v1.1.0) and the source from GitHub (v1.2.0). Unfortunately, I ran into several issues that prevented me from successfully running the package.

1. **Cython / Cubic_spline Issue**
   It seems that the PyPI version does not include the compiled Cython interface for `Cubic_spline`. This leads to the following error when importing ASTRA modules:

```text
CRITICAL | ASTRA.utils.Cubic_spline:<module>:17 - Cython interface is not found, please make sure that the installation went smoothly
```

2. **Multiprocessing / ASTRA.utils Import Issue**
When installing from the GitHub source and running DataClassManager().start(), ASTRA tries to start multiple processes in the background (multiprocessing). On macOS, these new processes cannot locate the ASTRA.utils module, resulting in:

```text
ModuleNotFoundError: No module named 'ASTRA.utils'
EOFError
```

3. **First Attempt with DACE ESPRESSO Files**
The first files I tried were downloaded from DACE (Data & Analysis Center for Exoplanets) for Epsilon Indi A: [DACE RV data](https://dace.unige.ch/radialVelocities/?pattern=eps%20indi#).

Using these files, I got the Cython issue (with the PyPI version), and I got another one:

```text
KeyError: "Keyword 'ESO QC CCF MASK' not found."
```

4. **ESPRESSO FITS Files Not Recognized (astroquery download)**
As I got the ESO MASK error, I tried with astroquery.
This gave me another issue, namely ESPRESSO FITS files were not recognized by the `DataClass`:

```text
Frame - ESPRESSO can't recognize the file that it received!
```

5. **General Notes / Recommendations**
- macOS appears to be sensitive to the multiprocessing and Cython issues. I don't know if you've ever run into this one.
- In the notebooks, it would be helpful to use placeholder paths instead of hard-coded local paths for reproducibility (that's another thing, but I'd like to mention it).
- I'm curious about the differences between PyPI v1.1.0 and GitHub v1.2.0.

**Summary:**
It can happen that I missed something, or that I have configuration issues. If you have encountered any of these problems yourself or if you have any suggestions on how to resolve them, I would appreciate your assistance. :)

I hope this feedback is useful.

Best regards,
Dóri
