### 2025-11-27 a711c5e
Setup of automatic pypi publish
