+++ .github/workflows/publishing.yml
@@ -1,0 +1,28 @@
+name: "Publish"
+
+on:
+  push:
+    tags:
+      # Publish on any tag starting with a `v`, e.g., v0.1.0
+      - v*
+
+jobs:
+  run:
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+    permissions:
+      id-token: write
+      contents: read
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v5
+      - name: Install uv
+        uses: astral-sh/setup-uv@v7
+      - name: Install Python 3.12
+        run: uv python install 3.12
+      - name: Build
+        run:  uv build
+      # Check that basic features work and we didn't miss to include crucial files
+      - name: Publish
+        run: uv publish
