### 2025-11-27 #9413 @Kamuish
Hey @tnehezd ,

Checking in again, regarding your comments:

## Multiprocessing / ASTRA.utils Import Issue

I was reading and I am fairly sure this is coming from the default method to spawn new processes through the multiprocessing package (different defaults in Linux/MacOS). Could you try the following, and let me know if it works? If so, I will update the docs to reflect that:

```
import multiprocessing as mp
mp.set_start_method('forkserver')
```

## ESPRESSO files

After I get the requested info, I will look into it

## Cython compilation

I am not able to reproduce the issue locally, with the pypi version (installed on a fresh environment). Could you let me know which pip version you are using?

## Pypi vs GH version

Automated the release process, so that the pypi version is up to date with the GH version
