### 2025-11-27 e023711
Update OBSstellar to not break with oversampling routines
