+++ src/ASTRA/__init__.py
@@ -1,6 +1,6 @@
 """ASTRA - interface for spectra."""
 
-version = "0.1.0"
+version = "1.2.1"
 
 __version__ = version.replace(".", "-")
 __version_info__ = (int(i) for i in __version__.split("-"))
