+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ASTRA-spectra"
-version = "1.2.0"
+version = "1.2.1"
 description = "Interface to open stellar spectra and construct stellar and telluric models"
 readme = "README.md"
 requires-python = ">=3.11,<=3.12.8"
