### 2025-11-29 #9413 @Kamuish
Hey @tnehezd ,

No problems with the delays :)

As for the file that you mentioned:

No, those files contain the telluric spectrum that was adjusted to the actual spectra! You should be using the files that end in any of the following formats:  S2D_A.fits, S2D_BLAZE_A.fits, S2D_TELLCORR_A.fits, S2D_TELLCORR_A.fits, S2D_BLAZE_TELL_CORR_A.fits   (these are the ones that contain the actual 2-dimensional spectra)

You should have those files in the data that you downloaded. If that is not the case, do let me know so that I can provide you with some
