### 2025-11-29 b10bef1
Add validity for frames from ESO pipelines
