### 2025-11-29 4348a13
add badges
