### 2025-12-04 #9413 @tnehezd
Hi @Kamuish,

As for the multiproccessing, I had no issue with pypi. Maybe a good solution could be something like this

import multiprocessing as mp

try:
    mp.set_start_method('forkserver')
except RuntimeError:
    pass

--- if it's set, it will just pass and give no runtime error if it's already set.

I'll come back with the followings.
