### 2025-12-04 #9413 @tnehezd
Hi @Kamuish,

Thanks a lot for the instructions — it looks like I was finally able to select the correct FITS files, and the pipeline reads them.

I successfully installed the ASTRA‑spectra package from PyPI, but the pipeline failed during initialization with the following error:

> TypeError: 'str' object is not callable

This comes from the line in ESO_PIPELINE.py:

> if key in self.file_path.stem():

I found that removing the parentheses fixes the issue:

> if key in self.file_path.stem

I tested this on the GitHub version, and it seems to be a working solution: the pipeline now loads the spectra correctly. For reference, the DataClass reports:

> Data Class from <class 'ASTRA.Instruments.ESPRESSO.ESPRESSO'> holding 0 OBS from ESPRESSO18, 3 OBS from ESPRESSO19

I have some outputs in the tmp/templates/Stellar/Iteration_0/ dir. However, after this point the telluric correction fails. The telluric model appears to be generated, but applying it always results in:

> BadTemplateError: Template was not created

This suggests that the template object exists, but the internal telluric model may be empty or incomplete.
To check whether this might be related to the specific spectra, could you please send me a few spectra? Maybe on a drive link, or so... The pipeline does load the existing files, but I’d like to verify whether the telluric issue is data‑dependent.

I'm working with these like: r.ESPRE.2021-04-30T08/43/44.577_S2D_A.fits, r.ESPRE.2021-04-30T08/43/44.577_S2D_BLAZE_A.fits, r.ESPRE.2021-04-30T08/43/44.577_S2D_BLAZE_TELL_CORR_A.fits from DACE.

I have some more spectra from other nights, I have the same issue with telluric model. Could you give me some hints?

The code I use is from the basic_example

```
from pathlib import Path

from ASTRA.data_objects import DataClassManager
from ASTRA.data_objects.DataClass import DataClass
from ASTRA.Instruments import ESPRESSO
from ASTRA.Quality_Control.activity_indicators import Indicators
from ASTRA.template_creation.StellarModel import StellarModel
from ASTRA.template_creation.TelluricModel import TelluricModel
from ASTRA.utils.choices import TELLURIC_CREATION_MODE

storage_path = Path("tmp")
manager = DataClassManager()
manager.start()

d: DataClass = manager.DataClass(
    list(Path("ASTRA/EPR19/2019-04-29").glob("*.fits")),
    storage_path=storage_path,
    instrument=ESPRESSO,
)

inds = Indicators()
d.remove_activity_lines(inds)

ModelTell = TelluricModel(
    usage_mode="individual",
    user_configs={"CREATION_MODE": TELLURIC_CREATION_MODE.telfit},
    root_folder_path=storage_path,
)

ModelTell.Generate_Model(
    dataClass=d,
    telluric_configs={},
    force_computation=False,
    store_templates=True,
)
d.remove_telluric_features(ModelTell)

stell = StellarModel(
    storage_path,
    user_configs={},
)
stell.Generate_Model(
    dataClass=d,
    template_configs={},
)
```

Best regards,
Dóri
