### 2025-12-04 45ecb32
Fix issue with initialization of ESO based pipelines in v1.2.2
