+++ src/ASTRA/Instruments/ESO_PIPELINE.py
@@ -149,7 +149,7 @@ class ESO_PIPELINE(Frame):
         if not self._internal_configs["use_old_pipeline"]:
             logger.debug("Checking for validity of file reduced with new pipeline")
             for key in ["S2D_A", "S2D_BLAZE_A", "S2D_BLAZE_TELL_CORR_A", "S2D_TELL_CORR_A", "S1D_A"]:
-                if key in self.file_path.stem():
+                if key in self.file_path.stem:
                     break
             else:
                 raise custom_exceptions.InvalidConfiguration(f"{self.name} can't recognize {self.file_path}")
