### 2025-12-04 #9413 @tnehezd
Hi @Kamuish,

This is the relevant part of the log:

```
 ...

CRITICAL | ASTRA.utils.custom_exceptions:inner1 - Template creation failed
ModuleNotFoundError: No module named 'telfit'
InternalError: Telfit is not installed
...

```

So it turned out that the telluric model fails simply because TELFIT is not installed. I tried installing TELFIT under Python 3.11, but it doesn’t build — even though Cython is installed. The installation fails during metadata generation with:

> AssertionError: Multiple .egg-info directories found
> error: metadata-generation-failed

At the moment, it seems I’m stuck in a version mismatch: TELFIT cannot be installed under Python 3.11, while ASTRA‑spectra does not run under Python 3.10. Because of this, I can’t get the TELFIT‑based telluric model to run.
Do you have any ideas how to solve it?

As a remark: maybe I missed it, but it would be helpful if the documentation mentioned that TELFIT needs to be installed separately. If this is a common issue, a short “Troubleshooting” section might also help future users.

Thanks,
Dóri
