### 2025-12-11 d9bcccf
Add missing S2D file formats
