### 2025-12-18 9c69450
Update of HARPS file to ensure fallback to default value to use_old_pipeline when not passed by user
