### 2025-12-19 5a95280
Update of log messages for templates that were not created
