+++ src/ASTRA/base_models/TemplateFramework.py
@@ -103,7 +103,7 @@ class TemplateFramework(BASE):
         """
         try:
             if not self.templates[subInstrument].is_valid:
-                msg = "Template was not created"
+                msg = "Template was not created, please refer to previous log messages for details"
                 logger.critical(msg)
                 raise custom_exceptions.BadTemplateError(msg)
         except KeyError:
