### 2025-12-22 8ef4238
Remove hard coded path for the tests
