### 2026-01-06 #9413 @Romain-Thomas-Shef
@Kamuish Can you make a release of the software that includes all the changes that have been made during the review? Once this is done you need to archive the software and provide me with the DOI. Various platform are available for this like zenodo or figshare.  While doing so, can you make sure that:

-The title of the archive is the same as the JOSS paper title
-That the authors of the archive are the same as the JOSS paper authors
