### 2026-01-09 #9413 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
