### 2026-01-09 bfbc2a1
Initial editorial changes for publication in JOSS
