+++ paper/paper.bib
@@ -62,7 +62,7 @@
 
 @inproceedings{maroonx_paper,
 author = {Andreas Seifahrt and Jacob L. Bean and David Kasper and Julian St{\"u}rmer and Madison Brady and Robert Liu and Mathias Zechmeister and Gudmundur K. Stef{\'a}nsson and Ben Montet and John White and Eduardo Tapia and Teo Mocnik and Siyi Xu and Christian Schwab},
-title = {{MAROON-X: the first two years of EPRVs from Gemini North}},
+title = {{MAROON-X}: the first two years of {EPRVs} from {Gemini North}},
 volume = {12184},
 booktitle = {Ground-based and Airborne Instrumentation for Astronomy IX},
 editor = {Christopher J. Evans and Julia J. Bryant and Kentaro Motohara},
@@ -112,7 +112,7 @@ URL = {https://doi.org/10.1117/12.2056453}
 }
 
 @article{artigau_linebyline2022,
-  title      = {Line-by-line {Velocity} {Measurements}: an {Outlier}-resistant {Method} for {Precision} {Velocimetry}},
+  title      = {Line-by-line Velocity Measurements: an Outlier-resistant Method for Precision Velocimetry},
   volume     = {164},
   issn       = {0004-6256, 1538-3881},
   shorttitle = {Line-by-line {Velocity} {Measurements}},
@@ -160,10 +160,10 @@ URL = {https://doi.org/10.1117/12.2056453}
 
 @article{Stangret2024,
 	author = {{Stangret, M.} and {Palle, E.} and {Esparza-Borges, E.} and {Orell Miquel, J.} and {Casasayas-Barris, N.} and {Zapatero Osorio, M. R.} and {Cristo, E.} and {Allart, R.} and {Alibert, Y.} and {Borsa, F.} and {Demangeon, O. D. S.} and {Di Marcantonio, P.} and {Ehrenreich, D.} and {Figueira, P.} and {González Hernández, J. I.} and {Herrero-Cisneros, E.} and {Martins, C. J. A. P.} and {Santos, N. C.} and {Seidel, J. V.} and {Azevedo Silva, T.} and {Sozzetti, A.} and {Steiner, M.} and {Suárez Mascareño, A.} and {Udry, S.}},
-	title = {The obliquity and atmosphere of the hot Jupiter WASP-122b (KELT-14b) with ESPRESSO: An aligned orbit and no sign of atomic or molecular absorption★},
+	title = {The obliquity and atmosphere of the hot {Jupiter} WASP-122b (KELT-14b) with ESPRESSO: An aligned orbit and no sign of atomic or molecular absorption⋆},
 	DOI= "10.1051/0004-6361/202450938",
 	url= "https://doi.org/10.1051/0004-6361/202450938",
-	journal = {A&A},
+	journal = {Astronomy \& Astrophysics},
 	year = 2024,
 	volume = 691,
 	pages = "A120",
@@ -171,10 +171,10 @@ URL = {https://doi.org/10.1117/12.2056453}
 
 @article{Damasceno2024,
[…]
