+++ paper/paper.md
@@ -1,34 +1,34 @@
 ---
-title: 'Astra: A Python Package for Cross-Instrument Stellar and Telluric Template Construction'
+title: 'ASTRA: A Python Package for Cross-Instrument Stellar and Telluric Template Construction'
 tags:
   - Python
   - astronomy
 authors:
   - name: André M. Silva
     orcid: 0000-0003-4920-738X
-    affiliation: "1, 2" 
+    affiliation: "1, 2"
   - name: J. P. Faria
     orcid:  0000-0002-6728-244X
-    affiliation: 3 
+    affiliation: 3
   - name: Nuno C. Santos
     orcid:  0000-0003-4422-2919
-    affiliation: "1, 2" 
+    affiliation: "1, 2"
   - name: Sérgio G. Sousa
     orcid:  0000-0001-9047-2965
-    affiliation: 1 
+    affiliation: 1
   - name: Pedro T. P. Viana
     orcid:  0000-0003-1572-8531
-    affiliation: "1, 2" 
+    affiliation: "1, 2"
   - name: J. H. C. Martins
     orcid:  0000-0002-1532-9082
-    affiliation: 1 
+    affiliation: 1
 
 affiliations:
  - name: Instituto de Astrofísica e Ciências do Espaço, CAUP, Universidade do Porto, Rua das Estrelas, 4150-762 Porto, Portugal
    index: 1
  - name: Departamento de Física e Astronomia, Faculdade de Ciências, Universidade do Porto, Rua do Campo Alegre, 4169-007 Porto, Portugal
    index: 2
- - name: Département d’astronomie de l’Université de Genève, Chemin Pegasi 51, 1290 Versoix, Switzerland 
+ - name: Département d’astronomie de l’Université de Genève, Chemin Pegasi 51, 1290 Versoix, Switzerland
    index: 3
 date: 23 April 2025
 bibliography: paper.bib
@@ -36,32 +36,30 @@ bibliography: paper.bib
 
 # Summary
 
-*ASTRA* is a Python package that provides a modular, instrument-independent interface for working with high-resolution stellar spectra. Designed to support data from multiple spectrographs — including ESPRESSO [@pepeESPRESSOVLTOnsky2021], HARPS [@2003MsngrHARPS; @pepe_harps_2002], MAROON-X [@maroonx_paper], and CARMENES [@carmenes_paper] — *ASTRA* offers a unified abstraction over their data formats, enabling consistent access to fluxes, wavelengths, uncertainties, and metadata across instruments. Furthermore, it applies the necessary wavelength and flux calibrations that are needed, as described by the official pipelines of each instrument.
+*ASTRA* is a Python package that provides a modular, instrument-independent interface for working with high-resolution stellar spectra. Designed to support data from multiple spectrographs—including ESPRESSO [@pepeESPRESSOVLTOnsky2021], HARPS [@2003MsngrHARPS; @pepe_harps_2002], MAROON-X [@maroonx_paper], and CARMENES [@carmenes_paper]—*ASTRA* offers a unified abstraction over their data formats, enabling consistent access to fluxes, wavelengths, uncertainties, and metadata across instruments. Furthermore, it applies the necessary wavelength and flux calibrations that are needed, as described by the official pipelines of each instrument.
 
 In addition to a common interface, *ASTRA* provides internal quality control checks of the observations, automatically divides them into the different sub-datasets that are commonly used by each spectrograph, and provides avenues to dynamically reject observations based on different properties. Furthermore, it also provides routines to mask the spectral imprint of Earth's atmosphere (in the form of telluric lines) and construct high-SNR, data-driven, stellar templates.
 
[…]
